+++ internal/parsers/parser_gcov/parser.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/IgorBayerl/AdlerCov/internal/filereader"
+	"github.com/IgorBayerl/AdlerCov/filereader"
 	"github.com/IgorBayerl/AdlerCov/internal/parsers"
 )
